# Conflicts:
#	src/common/platform/win32/i_dijoy.cpp
#	src/common/platform/win32/i_main.cpp
#	src/common/platform/win32/win32polyvideo.cpp
#	src/d_main.cpp
#	src/g_statusbar/shared_sbar.cpp
#	wadsrc/static/menudef.txt
#	wadsrc/static/zscript/engine/ui/menu/loadsavemenu.zs
#	wadsrc/static/zscript/engine/ui/menu/menu.zs
#	wadsrc/static/zscript/engine/ui/menu/optionmenu.zs

+++ src/playsim/p_acs.cpp
@@ -4248,11 +4248,11 @@ void DLevelScript::DoSetActorProperty (AActor *actor, int property, int value)
 		break;
 
 	case APROP_ScaleX:
-		actor->Scale.X = ACSToDouble(value);
+		actor->Scale.X = (float)ACSToDouble(value);
 		break;
 
 	case APROP_ScaleY:
-		actor->Scale.Y = ACSToDouble(value);
+		actor->Scale.Y = (float)ACSToDouble(value);
 		break;
 
 	case APROP_Mass:
